chore: move remaining unit tests to /src
