+++ react-components/src/hooks/use3dModels.test.ts
@@ -1,12 +1,15 @@
+/*!
+ * Copyright 2025 Cognite AS
+ */
 import { describe, expect, test, vi, beforeEach, beforeAll, afterAll } from 'vitest';
-import { use3dModels } from '../../../src/hooks/use3dModels';
 import { type CogniteCadModel, type CogniteModel } from '@cognite/reveal';
 import { renderHook } from '@testing-library/react';
 
-import { viewerMock, viewerModelsMock } from '../fixtures/viewer';
-import { cadMock } from '../fixtures/cadModel';
 import { Mock } from 'moq.ts';
 import { Matrix4 } from 'three';
+import { cadMock } from '../../tests/unit-tests/fixtures/cadModel';
+import { viewerMock, viewerModelsMock } from '../../tests/unit-tests/fixtures/viewer';
+import { use3dModels } from './use3dModels';
 
 const mockResourceCount = { reveal3DResourcesCount: 2 };
